Move RendererConnect to shared2
