+++ packages/react-cosmos-fixture/src/FixtureLoader/webSockets.ts
@@ -1,6 +1,8 @@
 import io from 'socket.io-client';
-import { RENDERER_MESSAGE_EVENT_NAME } from 'react-cosmos-shared2/renderer';
-import { RendererConnect } from '../shared';
+import {
+  RENDERER_MESSAGE_EVENT_NAME,
+  RendererConnect
+} from 'react-cosmos-shared2/renderer';
 
 export function createWebSocketsConnect(url: string): RendererConnect {
   const socket = io(url);
