+++ packages/react-cosmos-fixture/src/index.ts
@@ -5,7 +5,7 @@ export { createWebSocketsConnect } from './FixtureLoader/webSockets';
 
 // Use in decorators
 export { FixtureState } from 'react-cosmos-shared2/fixtureState';
-export { SetFixtureState, RendererConnect } from './shared';
+export { SetFixtureState } from './shared';
 export { FixtureContext } from './FixtureContext';
 
 // Advanced: Use in fixtures to capture elements in render callback
