+++ packages/react-cosmos/src/plugins/webpack/client/hmrErrorHandler.ts
@@ -1,6 +1,8 @@
-import { RendererRequest } from 'react-cosmos-shared2/renderer';
+import {
+  RendererRequest,
+  RendererConnect
+} from 'react-cosmos-shared2/renderer';
 import { WebpackRendererResponse } from 'react-cosmos-shared2/webpack';
-import { RendererConnect } from 'react-cosmos-fixture';
 import { getRendererId, getRendererConnect } from '../../../domRenderer';
 
 declare var __DEV__: boolean;
